Fix issue with formatting
When formatting typescript and tsx files there has been a bug where we
the last line gets repeated. While reading up and trying to find the
issue I noticed that null-ls is depricated. There is a community fork
that is being maintained called none-ls that I switched to. But that
didn't solve the issue. Instead I switched to using conform for
formatting, that solved it!